Move alternate lang to header and fix lint issues
